Rename varda packages and classes
This finishes the removal of old varda integrations.
